+++ skimage/_shared/transform.pxd
@@ -2,4 +2,4 @@ cimport numpy as cnp
 
 
 cdef float integrate(cnp.ndarray[float, ndim=2,  mode="c"] sat,
-                     int r0, int c0, int r1, int c1)
+                     ssize_t r0, ssize_t c0, ssize_t r1, ssize_t c1)
